Update verification events
